G15LCDEngine_helper: change size parameter in submitFrame() to qint64.
This helps make it compatible with QIODevice's write()
method without casting.

Since it's an internal method with only one consumer,
this seems like the most sensible choice.

+++ src/mumble/G15LCDEngine_helper.cpp
@@ -117,7 +117,7 @@ bool G15LCDEngineHelper::framebufferReady() const {
 	return !bUnavailable && (qpHelper->state() == QProcess::Running);
 }
 
-void G15LCDEngineHelper::submitFrame(bool alert, unsigned char *buf, size_t len) {
+void G15LCDEngineHelper::submitFrame(bool alert, unsigned char *buf, qint64 len) {
 	char pri = alert ? 1 : 0;
 	if ((qpHelper->write(&pri, 1) != 1) || (qpHelper->write(reinterpret_cast<char *>(buf), len) != len))
 		qWarning("G15LCDEngine_lglcd: failed to write");
